fixed Enum
